+++ config/crd/bases/core.telespazio-uk.io_workspaces.yaml
@@ -53,6 +53,8 @@ spec:
           status:
             description: WorkspaceStatus defines the observed state of Workspace
             properties:
+              awsRole:
+                type: string
               namespace:
                 description: Name of child namespace
                 type: string
