+++ templates/kubernetes/service-account.yaml
@@ -1,0 +1,5 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+    name: aws
+    namespace: {{.Workspace.Name}}
